requirements.txt
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
seaborn>=0.12.0
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.20.0
scikit-learn>=1.0.0
xgboost>=2.0.0
lightgbm>=4.0.0
scipy>=1.9.0
shap>=0.41.0
pyyaml>=6.0
pytest>=7.0.0

